+++ wfapi/__init__.py
@@ -14,7 +14,7 @@ You can add node, edit, complete, or uncomplete, etc.
 __project_url__ = "http://github.com/sigsrv/wfapi"
 __author__ = "sigsrv (sigsrv@sigsrv.net)"
 
-__version__ = "0.2.0-alpha"
+__version__ = "0.2.5-alpha"
 # based on (github commit count - 22) in 0.2.x
 # https://www.python.org/dev/peps/pep-0396/
 # http://semver.org/lang/ko/