+++ wfapi/hashtag.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+raise NotImplementedError
+
+TAG_SHARP = "#"
+TAG_AT = "@"
+
+# TODO: provide api for search.