+++ wfapi/search/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+raise NotImplementedError
+
+from .pattern import *
+from .searcher import *