+++ wfapi/search/pattern.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import re
+
+# TODO: support search by workflowy's keyword
+
+class WFBasePattern():
+    pass
+