+++ wfapi/shared.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 raise NotImplementedError
+import enum #?
+
 
 class WFBaseSharedInfo():
     pass